requirements.txt
requests==2.27.1
pandas==2.1.4
psycopg2-binary==2.9.9
numpy==1.26.2
SQLAlchemy==1.4.38